test: split shell-ops and code-ops into unit and integration tests
